Add delay for container creation
Problem:
5 nodes in nodes.csv, with 3 containers called "Container1" and 2 containers called "Container2".
Calling script creates 3 containers called "Container1" and 2 containers called "Container2" instead of detecting duplicates.

The above occurs when 2 or more nodes with the same Container name are processed in quick succession, the duplicity check on the second node returns null because NetXMS Server hasn't yet registered the previous container properly. Adding a short delay after a new container is created ensures that when the next node is processed and the container duplicity check occurs, any previous container created will be included in the filtering operation.